Refactor continues
