+++ src/main/java/conj/shop/addons/ShopFile.java
@@ -1,6 +1,6 @@
 package conj.shop.addons;
 
-import conj.shop.base.Initiate;
+import conj.shop.Initiate;
 import conj.shop.commands.control.Manager;
 import conj.shop.data.Page;
 import conj.shop.data.PageSlot;
