+++ src/main/java/conj/shop/data/enums/Config.java
@@ -2,7 +2,7 @@ package conj.shop.data.enums;
 
 import conj.shop.tools.auto.Autobackup;
 import conj.shop.tools.auto.Autosave;
-import conj.shop.base.Initiate;
+import conj.shop.Initiate;
 import org.bukkit.ChatColor;
 import org.bukkit.plugin.Plugin;
