+++ src/main/java/conj/shop/events/listeners/Input.java
@@ -1,6 +1,6 @@
 package conj.shop.events.listeners;
 
-import conj.shop.base.Initiate;
+import conj.shop.Initiate;
 import conj.shop.commands.control.Manager;
 import conj.shop.data.Page;
 import conj.shop.events.custom.PlayerInputEvent;
