+++ src/main/java/conj/shop/events/listeners/TradeEditor.java
@@ -1,6 +1,6 @@
 package conj.shop.events.listeners;
 
-import conj.shop.base.Initiate;
+import conj.shop.Initiate;
 import conj.shop.data.Page;
 import conj.shop.data.PageSlot;
 import conj.shop.data.enums.Function;
