+++ src/main/java/conj/shop/tools/auto/Autosave.java
@@ -1,6 +1,6 @@
 package conj.shop.tools.auto;
 
-import conj.shop.base.Initiate;
+import conj.shop.Initiate;
 import conj.shop.commands.control.Manager;
 import conj.shop.data.Page;
 import conj.shop.data.enums.Config;
